+++ packages/rulesets/src/oas/__tests__/oas3-operation-security-defined.test.ts
@@ -25,6 +25,28 @@ testRule('oas3-operation-security-defined', [
     },
     errors: [],
   },
+  {
+    name: 'validate a correct object (API-level security)',
+    document: {
+      openapi: '3.0.2',
+      components: {
+        securitySchemes: {
+          apikey: {},
+        },
+        security: [
+          {
+            apikey: [],
+          },
+        ],
+      },
+      paths: {
+        '/path': {
+          get: {},
+        },
+      },
+    },
+    errors: [],
+  },
 
   {
     name: 'return errors on invalid object',
@@ -46,7 +68,93 @@ testRule('oas3-operation-security-defined', [
     errors: [
       {
         message: 'Operation "security" values must match a scheme defined in the "components.securitySchemes" object.',
-        path: ['paths', '/path', 'get', 'security', '0'],
+        path: ['paths', '/path', 'get', 'security', '0', 'apikey'],
+        severity: DiagnosticSeverity.Warning,
+      },
+    ],
+  },
+
+  {
+    name: 'return errors on invalid object (API-level)',
+    document: {
+      openapi: '3.0.2',
+      components: {},
+      security: [
+        {
+          apikey: [],
+        },
+      ],
+      paths: {
+        '/path': {
+          get: {},
+        },
+      },
+    },
+    errors: [
+      {
+        message: 'API "security" values must match a scheme defined in the "components.securitySchemes" object.',
+        path: ['security', '0', 'apikey'],
+        severity: DiagnosticSeverity.Warning,
+      },
+    ],
+  },
+
+  {
+    name: 'return errors on valid and invalid object',
+    document: {
+      openapi: '3.0.2',
+      components: {
+        securitySchemes: {
+          apikey: {},
+        },
+      },
+      paths: {
+        '/path': {
+          get: {
+            security: [
+              {
+                apikey: [],
+                basic: [],
+              },
+            ],
+          },
+        },
+      },
+    },
+    errors: [
+      {
+        message: 'Operation "security" values must match a scheme defined in the "components.securitySchemes" object.',
+        path: ['paths', '/path', 'get', 'security', '0', 'basic'],
+        severity: DiagnosticSeverity.Warning,
+      },
+    ],
+  },
+
+  {
+    name: 'valid and invalid object (API-level security)',
+    document: {
+      openapi: '3.0.2',
+      components: {
+        securitySchemes: {
+          apikey: {},
+        },
+      },
+      security: [
+        {
+          apikey: [],
+          basic: [],
+        },
+      ],
+      paths: {
+        '/path': {
+          get: {},
+        },
+      },
+    },
+    errors: [
+      {
+        message: 'API "security" values must match a scheme defined in the "components.securitySchemes" object.',
+        path: ['security', '0', 'basic'],
         severity: DiagnosticSeverity.Warning,
       },
     ],