+++ packages/functions/CHANGELOG.md
@@ -1,3 +1,17 @@
+# [@stoplight/spectral-functions-v1.6.0](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-functions-v1.5.2...@stoplight/spectral-functions-v1.6.0) (2022-03-03)
+
+
+### Features
+
+* **cli:** improve error logging ([#2071](https://github.com/stoplightio/spectral/issues/2071)) ([b194368](https://github.com/stoplightio/spectral/commit/b194368164d92dce31b7ceba84ccc94fbe51f979))
+
+# [@stoplight/spectral-functions-v1.5.2](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-functions-v1.5.1...@stoplight/spectral-functions-v1.5.2) (2022-02-28)
+
+
+### Bug Fixes
+
+* **functions:** __importDefault undefined ([609ecb1](https://github.com/stoplightio/spectral/commit/609ecb1b23f354459f96687f27f911e915cb6ab3))
+
 # [@stoplight/spectral-functions-v1.5.1](https://github.com/stoplightio/spectral/compare/@stoplight/spectral-functions-v1.5.0...@stoplight/spectral-functions-v1.5.1) (2021-12-29)
 
 ### Bug Fixes